Restore nint() as it is wanted by Ricardo

+++ GEOS_Shared/GEOS_Utilities.F90
@@ -535,7 +535,7 @@
        end if
 
        TI = (TI - TMINTBL)*DEGSUBS+1
-       IT = int(TI)
+       IT = nint(TI)
 
        if(URAMP==TMIX) then
           DQ    = ESTBLX(IT+1) - ESTBLX(IT)
@@ -724,7 +724,7 @@
          end if
 
          TT = (TI - TMINTBL)*DEGSUBS+1
-         IT = int(TT)
+         IT = nint(TT)
 
          if(URAMP==TMIX) then
             DQQ =  ESTBLX(IT+1) - ESTBLX(IT)
